<dp commit>

+++ src/rot13/rot13.tests/Acceptance_tests.cs
@@ -15,12 +15,16 @@ namespace rot13.tests
 
         static void CopyFiles(string sourcePath, string destinationPath)
         {
-            Directory.Delete(destinationPath);
+            if (Directory.Exists(destinationPath))
+                Directory.Delete(destinationPath);
             
             var filenames = Directory.GetFiles(sourcePath, "*.*", SearchOption.AllDirectories);
             foreach (var fn in filenames) {
+                var destinationDirectory
+                Directory.CreateDirectory(Path.GetFullPath())
+                
                 var destFilename = fn.Replace(sourcePath, destinationPath);
-                File.Copy(fn, destFilename);
+                File.Copy(fn, destFilename, true);
             }
         }
     }
